+++ server/src/antlers/tags/coreTags.ts
@@ -110,7 +110,7 @@ import GetErrors from './core/getErrors/getErrors';
 import GetAllErrors from './core/getErrors/getAllErrors';
 import GetError from './core/getErrors/getError';
 import SetTag from './core/set';
-import MountTag from './core/mount';
+import MountUrlTag from './core/mountUrl';
 
 const coreTags = [
     If,
@@ -175,7 +175,7 @@ const coreTags = [
     LocalesCount,
 
     Mix,
-    MountTag,
+    MountUrlTag,
     NavTag,
     StructureTag,
     NavBreadcrumbs,